requirements.txt
azure-identity
azure-ai-projects
azure-ai-inference
